requirements.txt
sqlglot>=20.0
oracledb>=1.4
PyYAML>=6.0
